feat: library card component
